Move the `RouteAddRequest` method into a separate builder
For both the `RouteAddRequest` and `RouteDelRequest`, we need to build
a route message. However only the former provides helper methods to do
so.

Instead of duplicating code, create a separate builder, and use it for
both request types.
